pyproject.toml
[project]
name = "grafik"
version = "0.1.0"
description = "Modular layered image editor — Qwen-Image-Layered via fal.ai"
requires-python = ">=3.11"
dependencies = [
    "fal-client>=0.5",
    "Pillow>=10.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "streamlit>=1.35",
    "httpx>=0.27",
    "typer>=0.12",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "numpy>=1.24",
]

[project.optional-dependencies]
psd = ["psd-tools>=1.9"]
dev = ["pytest>=8.0"]

[project.scripts]
grafik = "grafik.cli.main:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["grafik*"]
